Move spectral_model into africanus.model.spectral
