--- africanus/model/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ["spectral_model"]
-
-from africanus.model.spec_model import spectral_model
