+++ africanus/model/spectral/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["spectral_model"]
+
+from africanus.model.spectral.spec_model import spectral_model
