+++ africanus/model/spectral/spec_model.py
